implementing sync methods - have to find out how to work with workspaces - set some TODO

+++ flux-file-watcher-codenvy-vfs/src/main/java/com/codenvy/flux/watcher/server/VFSRepository.java
@@ -10,6 +10,17 @@
  *******************************************************************************/
 package com.codenvy.flux.watcher.server;
 
+import static com.google.common.base.Preconditions.checkNotNull;
+
+import java.io.ByteArrayInputStream;
+import java.nio.file.Path;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.CopyOnWriteArraySet;
+
 import com.codenvy.api.project.server.ProjectService;
 import com.codenvy.api.project.shared.dto.ProjectDescriptor;
 import com.codenvy.flux.watcher.core.spi.RepositoryEvent;
@@ -20,87 +31,99 @@ import com.codenvy.flux.watcher.core.spi.Resource;
 import com.codenvy.flux.watcher.core.spi.Resource.ResourceType;
 import com.google.inject.Inject;
 
-import java.io.ByteArrayInputStream;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Set;
-import java.util.concurrent.CopyOnWriteArraySet;
-
-import static com.google.common.base.Preconditions.checkNotNull;
-
 /**
  * {@link com.codenvy.flux.watcher.core.spi.RepositoryProvider} implementation.
  *
  * @author Stéphane Tournié
  */
 public class VFSRepository implements RepositoryProvider {
+    
     private final Set<RepositoryListener> repositoryListeners;
     private final ProjectService          projectService;
+    private final ConcurrentMap<String, String> projects;
 
     @Inject
     public VFSRepository(ProjectService projectService) {
         this.repositoryListeners = new CopyOnWriteArraySet<>();
         this.projectService = projectService;
+        this.projects = new ConcurrentHashMap<>();
     }
 
     @Override
-    public boolean addProject(String projectId, String path) {
+    public boolean addProject(String projectId, String projectPath) {
         checkNotNull(projectId);
-        checkNotNull(path);
-        ProjectDescriptor project = null;
+        checkNotNull(projectPath);
+
+        ProjectDescriptor projectDesc = null;
         try {
-            // TODO set workspace & NewProject fields
-            project = projectService.createProject(null, projectId, null);
+            // TODO set workspace field
+            projectDesc = projectService.getProject(null, projectId);
         } catch (Exception e) {
             e.getMessage();
         }
-        return (project != null);
+        if (projectDesc != null) {
+            final String previousProjectPath = projects.putIfAbsent(projectId, projectPath);
+            if (previousProjectPath == null) {
+                return true;
+            }
+        }
+        return false;
     }
 
     @Override
     public boolean removeProject(String projectId) {
         checkNotNull(projectId);
-        try {
-            // TODO set workspace & path fields
-            projectService.delete(null, null);
-        } catch (Exception e) {
-            e.getMessage();
-        }
-        return true;
+        
+        final String projectPath = projects.remove(projectId);
+        return projectPath != null;
     }
 
     @Override
     public Resource getResource(String projectId, String path) {
         checkNotNull(projectId);
         checkNotNull(path);
+        
+        final String projectPath = projects.get(projectId);
+        if (projectPath != null) {
+            // TODO get file/folder from Codenvy VFS & return it as a Resource  
+            //projectService.getFile(null, path);
+        }
         return null;
     }
 
     @Override
     public void createResource(Resource resource) {
         checkNotNull(resource);
-        // TODO set workspace, parentPath & fileName field
-        String resourcePath = resource.path().substring(0);
-        String resourceName = resource.path().substring(0);
-        try {
-            if (resource.type() == ResourceType.FOLDER) {
-                projectService.createFile(null, resourcePath, resourceName, "folder", new ByteArrayInputStream(resource.content()));
-            } else {
-                projectService.createFile(null, resourcePath, resourceName, "file", new ByteArrayInputStream(resource.content()));
+        
+        final String projectPath = projects.get(resource.projectId());
+        if (projectPath != null) {
+            try {
+                // TODO set workspace, parentPath & fileName field
+                // TODO create a field 'name' in Resource?
+                if (resource.type() == ResourceType.FOLDER) {
+                    projectService.createFile(null, null, null, "folder", new ByteArrayInputStream(resource.content()));
+                } else {
+                    projectService.createFile(null, null, null, "file", new ByteArrayInputStream(resource.content()));
+                }
+            } catch (Exception e) {
+                e.getMessage();
             }
-        } catch (Exception e) {
-            e.getMessage();
         }
     }
 
     @Override
     public void deleteResource(Resource resource) {
         checkNotNull(resource);
-        try {
-            // TODO set workspace field
-            projectService.delete(null, resource.path());
-        } catch (Exception e) {
-            e.getMessage();
+        
+        final String projectPath = projects.get(resource.projectId());
+        if (projectPath != null) {
+            // TODO check that resource.path() exists in codenvy VFS
+            try {
+                // TODO set workspace field
+                projectService.delete(null, resource.path());
+            } catch (Exception e) {
+                e.getMessage();
+            }
         }
     }
